@@ -1,6 +1,6 @@
 {
 	"name": "flightdeck",
-	"version": "1.1.0",
+	"version": "1.2.0",
 	"private": true,
 	"engines": {
 		"node": ">=16.13",
@@ -22,8 +22,8 @@
 		"@types/react-dom": "^17.0.11",
 		"history": "^5.2.0",
 		"jwt-decode": "^3.1.2",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
+		"react": "^18",
+		"react-dom": "^18",
 		"react-router": "^6.2.1",
 		"react-router-dom": "^6.2.1",
 		"react-scripts": "^5.0.0",
