@@ -1,15 +1,19 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import {createRoot} from 'react-dom/client';
 import './css/app.css';
 import App from './App';
 // import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
-	<React.StrictMode>
-		<App/>
-	</React.StrictMode>,
-	document.getElementById('root')
-);
+const container = document.getElementById('app');
+const root = createRoot(container);
+root.render(<App/>);
+
+// ReactDOM.render(
+// 	<React.StrictMode>
+// 		<App/>
+// 	</React.StrictMode>,
+// 	document.getElementById('root')
+// );
 
 // // If you want to start measuring performance in your app, pass a function
 // // to log results (for example: reportWebVitals(console.log))
